pmrem: environment lighting, and the generator gated against three's numbers
`EnvironmentNode` wired into the physical material behind an optional
`pmrem_env` handle, so a material with an `envMap` resolves through the
cube-UV nodes. The two `addAssign`s land between `indirectDiffuse()` and
`indirectSpecular()`, which is where three's lighting-node ordering puts
them, and with them the declarations of `radiance` and `iblIrradiance`;
with no environment `indirectSpecular` declares them itself, so every
existing row of the ladder is byte-identical.

The 86-line radiance/iblIrradiance block matches `m07` statement for
statement. `m07` as a whole is r186's restructured `PhysicalLightingModel`,
which is a rung of its own; noted in `docs/nodes.md` §12.

`tests/pmrem.rs` gates the generator's numbers against oracles taken by
running three's own code: a real `PMREMGenerator` driven through
`_applyGGXFilter` with `_setViewport` and `render` replaced by recorders,
and `_generateCubeUVSize`. That found two underflows: `mipInt` is
`_lodMax - lodIn` in JS arithmetic and goes negative (-1, -2) for the extra
LODs, so both subtractions now happen in floating point.

+++ src/renderer/pmrem.rs
@@ -264,7 +264,9 @@ impl PmremGenerator {
             .expect("three-rs: _init() built the GGX material");
         uniforms.env_map.set_gpu(gpu_texture_of(&ping_pong));
         uniforms.roughness.set(vec![0.0]);
-        uniforms.mip_int.set(vec![(self.lod_max - lod_out) as f64]);
+        uniforms
+            .mip_int
+            .set(vec![self.lod_max as f64 - lod_out as f64]);
         set_viewport(target, x, y, width, height);
         renderer.set_render_target(Some(target.clone()));
         renderer.render_pmrem_mesh(geometry, &material, false);
@@ -310,7 +312,9 @@ pub fn ggx_step(lod_max: usize, lod_count: usize, lod_in: usize, lod_out: usize)
     // "Apply blur strength mapping for better quality across the roughness
     // range" — not a normalisation, a deliberate extra 1.25× of blur.
     let blur_strength = target_roughness * 1.25;
-    (incremental * blur_strength, (lod_max - lod_in) as f64)
+    // `_lodMax - lodIn` in JS arithmetic: `lodIn` runs past `lodMax` for the
+    // extra LODs, so the two go negative (-1 and -2 for a 256² source).
+    (incremental * blur_strength, lod_max as f64 - lod_in as f64)
 }
 
 /// The `( x, y )` of level `lod_out`'s tile and the level's face size.
